Fix - changing points field in ProductWithStockStatus to Double

+++ src/main/java/pl/grizzlysoftware/dotykacka/client/v1/api/dto/product/ProductWithStockStatus.java
@@ -88,7 +88,7 @@ public class ProductWithStockStatus {
     @JsonProperty("plu")
     public String plu;
     @JsonProperty("points")
-    public Integer points;
+    public Double points;
     @JsonProperty("priceVatCsv")
     public double priceVatCsv;
     @JsonProperty("pricewithoutvat")
